ref(nav) move isLinkActive and ProjectIcon inside components
